project/agent_server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
greenlet==3.1.1
